refactor(fase6): extract transformApiError and make errorHandler references safe; add tests
